lambdas/crypto-potential-ema-crossover-alert/requirements.txt
boto3
pandas
numpy==1.26.4
python-binance
requests
